Moved fetchGetURLs() from ContractResultAnalyzer to MockUtils.
