+++ tests/unit/MockUtils.ts
@@ -1,0 +1,38 @@
+/*-
+ *
+ * Hedera Mirror Node Explorer
+ *
+ * Copyright (C) 2021 - 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+
+//
+// Tools
+//
+
+import MockAdapter from "axios-mock-adapter";
+
+export function cloneMock(mock: object): any {
+    return JSON.parse(JSON.stringify(mock))
+}
+
+export function fetchGetURLs(mock: MockAdapter): string[] {
+    const result: string[] = []
+    for (const e of mock.history.get) {
+        result.push(e.url)
+    }
+    return result
+}
